Reorganization, allowing load settings from json
