+++ src/Settings/default_mappings.json
@@ -1,0 +1,17 @@
+[
+  {
+    "gesture": "thumbs_up",
+    "function": "volume_up",
+    "hexkey": "0xAF"
+  },
+  {
+    "gesture": "thumbs_down",
+    "function": "volume_down",
+    "hexkey": "0xAE"
+  },
+  {
+    "gesture": "fist",
+    "function": "play/pause",
+    "hexkey": "0xB3"
+  }
+]
