+++ src/run_program.py
@@ -1,4 +1,4 @@
-from test import GestureDetector
+from src.Gestures.gesture_detector import GestureDetector
 
 classifier = GestureDetector()
 classifier.run_backend_mainloop()
